[hail] Add IndexReadRDD class
Try to not serialize partFile or interval arrays to the workers.
